requirements.txt
#pip install -r requirements.txt
pyyaml
pylxd

